Fixes for win32

+++ src/inet_pton.c
@@ -92,7 +92,7 @@ inet_pton4(src, dst)
 		const char *pch;
 
 		if ((pch = (char *) strchr(digits, ch)) != NULL) {
-			uint new = *tp * 10 + (pch - digits);
+			uint8_t new = *tp * 10 + (pch - digits);
 
 			if (new > 255)
 				return (0);
@@ -140,7 +140,7 @@ inet_pton6(src, dst)
 	u_char tmp[IN6ADDRSZ], *tp, *endp, *colonp;
 	const char *xdigits, *curtok;
 	int ch, saw_xdigit;
-	uint val;
+	unsigned val;
 
 	memset((tp = tmp), 0, IN6ADDRSZ);
 	endp = tp + IN6ADDRSZ;
